Rebuild checkScenario func
Function was optimized for checking only the rows and cols that were
affected by last choice (hence x, y arguments). However when
implementing minimax that overload of arguments was making code very
unreadable. Now it's a bit slower because it has to go through the whole
array so O(n^2). That's a kind of trade-off I had to make. Since this is
just my personal project... right now Im making this commit so I can
focus on other things. Soon I will refactor the code.