Fix a bug with two moves at once
There was a bug when AI generated its initial move as 'o' player.
The "currentMarker" function was faulty causing the AI player to make
two moves in a "at once".

Problem:
What exactly was causing a problem was the way currentMarker worked.
It was getting the current marker based on the markers array, which
is absolutely incorrect. The "markers" array is meant only for
the game initialization stage (which is why I moved it to to the startGame
function). The marker of the current player should be acquired from
the players array based on the round number.